requirements.txt
Flask
requests
Werkzeug
MarkupSafe
itsdangerous
click
certifi
charset-normalizer
idna
urllib3
Jinja2
python-dotenv
python-docx
reportlab
openai
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
